Refactored and simplified facility filtering and pagination concerns
